backend/pyproject.toml
[tool.poetry]
name = "iwasbored-backend"
version = "0.1.0"
description = "Backend service for IWasBored prediction market platform"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
web3 = "^6.11.0"
aiorwlock = "^1.0.1"
tenacity = "^8.2.3"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
loguru = "^0.7.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.7.0"
flake8 = "^6.0.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "backend:start"
start-dev = "backend:start_dev"
start-prod = "backend:start_prod"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true